Reorganize file and folder structure to accomodate the new tokenizer
