+++ server/routes/api/demo.js
@@ -1,13 +1,11 @@
 const express = require("express");
 const router = express.Router();
 const demoCtrl = require("../../controllers/api/demo.js");
-const tokenizer = require("../../controllers/api/tokenizer/tokenizer.js");
 const { addText } = require("../../controllers/api/demo.js");
 const { verifyJWT } = require("./../../utils/jwt");
 
 router.get("/demo", demoCtrl.getDemo);
 router.post("/translate", demoCtrl.translateSentence);
 router.post("/tokenize", demoCtrl.tokenizeText);
-router.post("/tokenizer", tokenizer.tokenizeText);
 router.post('/texts', verifyJWT, addText);
 module.exports = router;
